Package Reference

Abstract Types

The following abstract types are used to derive concrete state and event types.

HSM.AbstractHsmEvent — Type

AbstractHsmEvent

Abstract HSM event type.

Concrete events must inherit AbstractHsmEvent.

HSM.AbstractHsmState — Type

AbstractHsmState

Abstract HSM state type.

Concrete states must inherit AbstractHsmState, and either contain an HsmStateInfo struct called state_info, or implement the parent_state(), parent_state!(), active_substate(), and active_substate!() getters and setters.

Pass the parent state to the HsmStateInfo constructor in the concrete state's constructor if using HsmStateInfo.

Example:

struct MyState <: AbstractHsmState
    state_info::HsmStateInfo
    app_info::MyStatefulApplicationInfo

    MyState(parent_state, app_info) = new(HSM.HsmStateInfo(parent_state), app_info)
end

Otherwise, call parent_state(obj, <parent state>) and active_substate(obj, nothing) when constructing your concrete state.

HSM.HsmStateInfo — Type

HsmStateInfo

State information needed by HSM interfaces that act on AbstractHsmStates.

parent_state::AbstractHsmState – initialized to the parent state of the concrete state, or nothing if it is the root state machine state.

active_substate::AbstractHsmState – initialized to nothing.

Public Interfaces

The following HSM interfaces are used in state machine implementations.

HSM.handle_event! — Method

handle_event!(state_machine::AbstractHsmState, event::AbstractHsmEvent)

Pass event to state_machine for processing.

This is the public interface to use to pass an event to a state machine for processing.

HSM.transition_to_deep_history! — Method

transition_to_deep_history(state_machine::AbstractHsmState, state::AbstractHsmState)

Change the active state of state_machine, following the active state of state as far as it goes.

This is the interface to use in a state's event handler (on_event!()) to transition to the deep history of a state.

struct Machine <: HSM.AbstractHsmState ; end
struct MyState <: HSM.AbstractHsmState ; end
struct MyOtherState <: HSM.AbstractHsmState ; end
struct MySubState1 <: HSM.AbstractHsmState ; end
struct MySubState2 <: HSM.AbstractHsmEvent ; end
struct MySubSubState1 <: HSM.AbstractHsmEvent ; end
struct MySubSubState2 <: HSM.AbstractHsmEvent ; end
struct MyEvent <: HSM.AbstractHsmEvent ; end

machine = Machine(nothing)
my_state = MyState(machine)
my_other_state = MyOtherState(machine)
my_sub_state_1 = MySubState1(my_other_state)
my_sub_state_2 = MySubState2(my_other_state)
my_sub_sub-state_1 = MySubSubState1(my_sub_state_1)
my_sub_sub-state_2 = MySubSubState2(my_sub_state_1)

function on_event!(state::MyState, event::MyEvent)
    # Will transition last active state in my_other_state -- my_other_state,
    # my_sub_state_1, my_sub-state_2, my_sub_sub_state_1, or my_sub_sub_state_2.
    transition_to_shallow_history!(machine, my_other_state)
end

This interface corresponds to an arrow to a deep history marker in a state.

@startuml
state Machine {
    state MyState
    state MyOtherState {
        state MySubState1 {
            state MySubSubState1
            state MySubSubState2
        }
        state MySubState2
    }

    MyState --> MyOtherState[H*]
}
@enduml

HSM.transition_to_shallow_history! — Method

transition_to_shallow_history!(state_machine::AbstractHsmState, state::AbstractHsmState)

Change the active state of state_machine, following the active state of state as much as one layer.

This is the interface to use in a state's event handler (on_event!()) to transition to the history of a state.

struct Machine <: HSM.AbstractHsmState ; end
struct MyState <: HSM.AbstractHsmState ; end
struct MyOtherState <: HSM.AbstractHsmState ; end
struct MySubState1 <: HSM.AbstractHsmState ; end
struct MySubState2 <: HSM.AbstractHsmEvent ; end
struct MyEvent <: HSM.AbstractHsmEvent ; end

machine = Machine(nothing)
my_state = MyState(machine)
my_other_state = MyOtherState(machine)
my_sub_state_1 = MySubState1(my_other_state)
my_sub_state_2 = MySubState2(my_other_state)

function on_event!(state::MyState, event::MyEvent)
    # Will transition to last active state in my_other_state -- my_other_state,
    # my_sub_state_1, or my_sub_state_2.
    transition_to_shallow_history!(machine, my_other_state)
end

This interface corresponds to an arrow to a history marker in a state.

@startuml
state Machine {
    state MyState
    state MyOtherState {
        state MySubState1
        state MySubState2
    }

    MyState --> MyOtherState[H]
}
@enduml

HSM.transition_to_state! — Method

transition_to_state!(machine::AbstractHsmState, state::AbstractHsmState)

Change the active state of state_machine to state.

This is the interface to use in a state's event handler (on_event!(), on_initialize()) to transition states.

struct Machine <: HSM.AbstractHsmState ; end
struct MyState <: HSM.AbstractHsmState ; end
struct MyOtherState <: HSM.AbstractHsmState ; end
struct MyEvent <: HSM.AbstractHsmEvent ; end

machine = Machine(nothing)
my_state = MyState(machine)
my_other_state = MyOtherState(machine)

function on_event!(state::MyState, event::MyEvent)
    # Active state becomes my_other_state.
    transition_to_state!(machine, my_other_state)
end

This interface corresponds to arrows from one state to another in a UML state chart.

@startuml
state Machine {
    state MyState
    state MyOtherState

    MyState --> MyOtherState
}
@enduml

Extendable Interfaces

The following interfaces are extended to implement event handlers for specific states and events.

HSM.on_entry! — Method

on_entry!(state::AbstractHsmState)

Global state entry handler.

Create entry handlers for your derived states when you want that state to execute something whenever it is entered. This includes transitions to itself.

struct MyState <: HSM.AbstractHsmState ; end

function on_entry!(state::MyState)
    # do something.
end

State-specific instances of on_entry!() correspond to "on entry / action" labels in state blocks in UML state charts.

@startuml
state Machine {
    state MyState
    MyState : on entry / do something
}
@enduml

The global state entry handler is the default behavior of a state. It does not do anything.

HSM.on_event! — Method

on_event!(state::AbstractHsmState, event::AbstractHsmEvent)

Global event handler.

Create event handlers for your derived states and events to handle those specific events in those states.

struct MyState <: HSM.AbstractHsmState ; end
struct MyEvent <: HSM.AbstractHsmEvent ; end

function on_event!(state::MyState, event::MyEvent)
    # process event.
    return true
end

State-specific event handlers are expected to return true when they handle an event, unless they specifically want their parent state to process the event as well.

State-specific instances of on_event!() correspond to "event / action" labels in state blocks or on state transitions in UML state charts.

@startuml
state ExampleMachine {
    state MyState 
    state MyOtherState

    MyState --> MyOtherState : MyEvent / do something
    MyOtherState : MyEvent / do something
}
@enduml

The global event handler is called as a last resort, and is expected to return false to generate an HsmUnhandledEventError.

HSM.on_exit! — Method

on_entry!(state::AbstractHsmState)

Global state exit handler.

Create exit handlers for your derived states when you want that state to execute something whenever it is exited. This includes transitions to itself.

struct MyState <: HSM.AbstractHsmState ; end

function on_exit!(state::MyState)
    # do something.
end

State-specific instances of on_exit!() correspond to "on exit / action" labels in state blocks in UML state charts.

@startuml
state Machine {
    state MyState
    MyState : on exit / do something
}
@enduml

The global state exit handler is the default behavior of a state. It does not do anything.

HSM.on_initialize! — Method

on_initialize!(state::AbstractHsmState)

Global state initializer.

Create initialiers for states that must be initialized upon entry. Generally, this is used when a state has sub-states, where one must be transitioned to when the state is entered. For example, the root state machine.

struct Machine <: HSM.AbstractHsmState ; end
struct MyState <: HSM.AbstractHsmState ; end

machine = Machine(nothing)
my_state = MyState(machine)

function on_initialize!(state::Machine)
    transition_to_state!(my_state)
end

State-specific instances of on_initialize!() correspond to the starting point trasitions in UML state charts.

@startuml
state Machine {
    state MyState

    [*] --> MyState
}
@enduml

The global state initializer is the default behavior of a state. It does not do anything.

Extendable Getters and Setters

The following getters and setters are extended to implement the required parent and active state requirements for concrete implementations of AbstractHsmState. Extending the getters and setters is only necessary if the concrete implementation does not have a state_info::HsmStateInfo member.

HSM.active_substate — Method

active_substate(obj::AbstracHsmState)

active_substate getter.

Extend this for your concrete class if your concrete struct does not include state_info::HsmStateInfo.

HSM.active_substate! — Method

active_substate!(obj::AbstractHsmState, value::Union{AbstractHsmState, Nothing})

active_substate setter.

Extend this for your concrete class if your concrete struct does not include state_info::HsmStateInfo.

HSM.parent_state — Method

parent_state(obj::AbstractHsmState)

parent_state getter.

Extend this for your concrete class if your concrete struct does not include state_info::HsmStateInfo.

HSM.parent_state! — Method

parent_state!(obj::AbstractHsmState, value::Union{AbstractHsmState, Nothing})

parent_state setter.

Extend this for your concrete class if your concrete struct does not include state_info::HsmStateInfo.

Exceptions

The following exceptions are defined and used by HSM.

HSM.HsmUnhandledEventError — Type

Unhandled event exception.

This gets thrown if an event is not handled by the state machine when handle_event!() is called. This usually means the root state machine state does not have an event handler for the event type, or is not returning true.

HSM.HsmStateTransitionError — Type

State transition failed exception.

This gets thrown if the destination state is not a part of the state machine when transition_to_state!() is called.